refactor Stackblitz component
